+++ src/benchmarks/benchmarks.spec.ts
@@ -18,7 +18,7 @@ describe('benchmarks', () => {
         // Find gas used
         let gasUsed = tracker.collect().reduce((a, v) => a + v.events.reduce((c, d) => d.$type === 'processed' ? c + d.gasUsed : c, 0n), 0n);
         expect(gasUsed).toMatchInlineSnapshot(`2612n`);
-        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`261`);
+        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`280`);
     });
     it('benchmark functions(inline)', async () => {
 
@@ -33,6 +33,6 @@ describe('benchmarks', () => {
         // Find gas used
         let gasUsed = tracker.collect().reduce((a, v) => a + v.events.reduce((c, d) => d.$type === 'processed' ? c + d.gasUsed : c, 0n), 0n);
         expect(gasUsed).toMatchInlineSnapshot(`2481n`);
-        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`256`);
+        expect(functions.init!.code.toBoc().length).toMatchInlineSnapshot(`275`);
     });
 });